Split part of do_add_tag_to_task out as do_add_tag_to_task_by_id.
